pyproject.toml
[project]
name = "ecg-visualizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "PyQt6",
    "pyserial",
    "pyqtgraph",
    "numpy"
]
